+++ x/distribution/types/keys.go
@@ -1,6 +1,7 @@
 package types
 
 import (
+	"cosmossdk.io/collections"
 	"encoding/binary"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -53,7 +54,7 @@ var (
 	ValidatorAccumulatedCommissionPrefix = []byte{0x07} // key for accumulated validator commission
 	ValidatorSlashEventPrefix            = []byte{0x08} // key for validator slash fraction
 
-	ParamsKey = []byte{0x09} // key for distribution module params
+	ParamsKey = collections.NewPrefix(9) // key for distribution module params
 )
 
 // GetValidatorOutstandingRewardsAddress creates an address from a validator's outstanding rewards key.