+++ CHANGELOG.md
@@ -211,6 +211,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
   * `simulation.NewOperationMsg` is now 2-arity instead of 3-arity with the obsolete argument `codec.ProtoCodec` removed.
   * The field `OperationMsg.Msg` is now of type `[]byte` instead of `json.RawMessage`.
 
+* (x/distribution) Use collections for params state management
+
 
 ### Client Breaking Changes